# cmd/api/docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: sushilawasthi2999@gmail.com
    name: Sushil Awasthi
  description: Backend API for the PlantGo system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://wtfisterms.com/terms/
  title: PlantGo API
  version: "1.0"
paths:
...
  /scan/image:
    post:
      consumes:
      - multipart/form-data
      description: Accepts real-time image uploads for scanning
      parameters:
      - description: Image to scan
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Process scanned image
      tags:
      - Scanner
...
swagger: "2.0"